docs/irc-whatwg/2013-09-24.html

<!doctype html>
<head>
  <title>#whatwg on 2013-09-24</title>
  <link rel="stylesheet" href="../style.css">
  <script src="../logs.js"></script>
</head>
<body><div class="wrapper">
<div class="sidebar">
<div class="title">#whatwg<br>2013-09-24<br><a href="plaintext/">plaintext logs</a></div>
<div class="nav">
<a href="2013-09-23" class="nav-link"><span>prev</span></a>
<a href="2013-09-25" class="nav-link"><span style="float:right">next</span></a>
</div>
    
<div class="room-list-wrapper">Active:<br>
<ul class="room-list">
<li><a href="../TC39_Delegates/">TC39 Delegates</a></li>
<li><a href="../TC39_Editors/">TC39 Editors</a></li>
<li><a href="../TC39_General/">TC39 General</a></li>
<li><a href="../WHATWG/">WHATWG</a></li>
</ul>
</div>
<br>
<div class="room-list-wrapper">Historical:<br>
<ul class="room-list">
<li><a href="../irc-tc39/">#tc39</a></li>
<li><a href="../irc-tc39-delegates/">#tc39-delegates</a></li>
<li><a href="../irc-tc39-editor-group/">#tc39-editor-group</a></li>
<li><a href="../irc-whatwg/" class="current-room">#whatwg</a></li>
</ul>
</div>
<div class="footer"><a href="https://github.com/bakkot/matrix-archive-bot">source on github</a></div>
</div>
<div class="rhs">
<div class="rhs-header">
<span id="error" style="color: red; display:none">error</span>
<input type="text" id="query" size=25 placeholder="Search #whatwg">
<a id="search-submit" class="button icon-link" title="Search">
  <svg class="icon" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512"><path d="M505 442.7L405.3 343c-4.5-4.5-10.6-7-17-7H372c27.6-35.3 44-79.7 44-128C416 93.1 322.9 0 208 0S0 93.1 0 208s93.1 208 208 208c48.3 0 92.7-16.4 128-44v16.3c0 6.4 2.5 12.5 7 17l99.7 99.7c9.4 9.4 24.6 9.4 33.9 0l28.3-28.3c9.4-9.4 9.4-24.6.1-34zM208 336c-70.7 0-128-57.2-128-128 0-70.7 57.2-128 128-128 70.7 0 128 57.2 128 128 0 70.7-57.2 128-128 128z"></path></svg>
</a>
</div>

<div class="log"><table id="log-table"><tbody id="log-tbody">

  <tr class="msg" id="L0"><td class="ts-cell"><a class="ts" href="#L0" alt="Mon Sep 23 2013 17:03:22 GMT-0700 (Pacific Daylight Time)">00:03</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=Hixie_@irc>Hixie_</span>&gt;</div></td><td class="msg-cell">jamesr__: back</td></tr>
  <tr class="msg" id="L1"><td class="ts-cell"><a class="ts" href="#L1" alt="Mon Sep 23 2013 17:03:42 GMT-0700 (Pacific Daylight Time)">00:03</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=Hixie_@irc>Hixie_</span>&gt;</div></td><td class="msg-cell">jamesr__: you still around?</td></tr>
  <tr class="msg" id="L2"><td class="ts-cell"><a class="ts" href="#L2" alt="Mon Sep 23 2013 17:04:05 GMT-0700 (Pacific Daylight Time)">00:04</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=Hixie_@irc>Hixie_</span>&gt;</div></td><td class="msg-cell">gsnedders is gone, bummer.</td></tr>
  <tr class="msg" id="L3"><td class="ts-cell"><a class="ts" href="#L3" alt="Mon Sep 23 2013 17:04:40 GMT-0700 (Pacific Daylight Time)">00:04</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=Hixie_@irc>Hixie_</span>&gt;</div></td><td class="msg-cell">there can be many event loops. they run in parallel. for example, if you have firefox and chrome both running on your machine, you&#039;ll have at least two event loops.</td></tr>
  <tr class="msg" id="L4"><td class="ts-cell"><a class="ts" href="#L4" alt="Mon Sep 23 2013 17:04:44 GMT-0700 (Pacific Daylight Time)">00:04</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=Hixie_@irc>Hixie_</span>&gt;</div></td><td class="msg-cell">also each worker has its own event loop.</td></tr>
  <tr class="msg" id="L5"><td class="ts-cell"><a class="ts" href="#L5" alt="Mon Sep 23 2013 17:04:57 GMT-0700 (Pacific Daylight Time)">00:04</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=Hixie_@irc>Hixie_</span>&gt;</div></td><td class="msg-cell">multiprocess event loops have one loop per process.</td></tr>
  <tr class="msg" id="L6"><td class="ts-cell"><a class="ts" href="#L6" alt="Mon Sep 23 2013 17:04:58 GMT-0700 (Pacific Daylight Time)">00:04</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=Hixie_@irc>Hixie_</span>&gt;</div></td><td class="msg-cell">etc</td></tr>
  <tr class="msg" id="L7"><td class="ts-cell"><a class="ts" href="#L7" alt="Mon Sep 23 2013 17:06:14 GMT-0700 (Pacific Daylight Time)">00:06</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=Hixie_@irc>Hixie_</span>&gt;</div></td><td class="msg-cell">gsnedders: see logs</td></tr>
  <tr class="msg" id="L8"><td class="ts-cell"><a class="ts" href="#L8" alt="Mon Sep 23 2013 17:52:37 GMT-0700 (Pacific Daylight Time)">00:52</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=leo_33@irc>leo_33</span>&gt;</div></td><td class="msg-cell"> in your opinion why would a headhunter continue sending you info about a job after a  job interview?</td></tr>
  <tr class="msg" id="L9"><td class="ts-cell"><a class="ts" href="#L9" alt="Mon Sep 23 2013 17:55:00 GMT-0700 (Pacific Daylight Time)">00:55</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=jamesr__@irc>jamesr__</span>&gt;</div></td><td class="msg-cell">Hixie_: i&#039;m around, but not for super long</td></tr>
  <tr class="msg" id="L10"><td class="ts-cell"><a class="ts" href="#L10" alt="Mon Sep 23 2013 17:55:17 GMT-0700 (Pacific Daylight Time)">00:55</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=jamesr__@irc>jamesr__</span>&gt;</div></td><td class="msg-cell">Hixie_: i got distracted by finding a 3-month old crash bug in our parser</td></tr>
  <tr class="msg" id="L11"><td class="ts-cell"><a class="ts" href="#L11" alt="Mon Sep 23 2013 17:55:37 GMT-0700 (Pacific Daylight Time)">00:55</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=leo_33@irc>leo_33</span>&gt;</div></td><td class="msg-cell"> in your opinion why would a headhunter continue sending you info about a job after a  job interview?</td></tr>
  <tr class="msg" id="L12"><td class="ts-cell"><a class="ts" href="#L12" alt="Mon Sep 23 2013 17:57:17 GMT-0700 (Pacific Daylight Time)">00:57</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=jamesr__@irc>jamesr__</span>&gt;</div></td><td class="msg-cell">Hixie_: if you had &lt;script&gt;document.write(&#039;&lt;link ...&gt;&lt;script src=...&gt;&lt;\/script&gt;&#039;);&lt;/script&gt;, wouldn&#039;t the nesting level be non-zero when parsing the inner &lt;/script&gt;, thus bypassing that section of the algorithm?</td></tr>
  <tr class="msg" id="L13"><td class="ts-cell"><a class="ts" href="#L13" alt="Mon Sep 23 2013 18:20:09 GMT-0700 (Pacific Daylight Time)">01:20</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=Hixie_@irc>Hixie_</span>&gt;</div></td><td class="msg-cell">jamesr__: hm.</td></tr>
  <tr class="msg" id="L14"><td class="ts-cell"><a class="ts" href="#L14" alt="Mon Sep 23 2013 18:20:33 GMT-0700 (Pacific Daylight Time)">01:20</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=Hixie_@irc>Hixie_</span>&gt;</div></td><td class="msg-cell">jamesr__: no</td></tr>
  <tr class="msg" id="L15"><td class="ts-cell"><a class="ts" href="#L15" alt="Mon Sep 23 2013 18:21:25 GMT-0700 (Pacific Daylight Time)">01:21</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=Hixie_@irc>Hixie_</span>&gt;</div></td><td class="msg-cell">jamesr__: it&#039;s 1 while the document.write() runs, but the inner script doesn&#039;t execute until the outer script has ended, when it&#039;s zero again.</td></tr>
  <tr class="msg" id="L16"><td class="ts-cell"><a class="ts" href="#L16" alt="Tue Sep 24 2013 01:01:17 GMT-0700 (Pacific Daylight Time)">08:01</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-5" title=hallvors@irc>hallvors</span>&gt;</div></td><td class="msg-cell">jgraham: yt? time for a testharness.js question?</td></tr>
  <tr class="msg" id="L17"><td class="ts-cell"><a class="ts" href="#L17" alt="Tue Sep 24 2013 01:08:17 GMT-0700 (Pacific Daylight Time)">08:08</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-3" title=darobin@irc>darobin</span>&gt;</div></td><td class="msg-cell">hallvors: you&#039;re doing it wrong</td></tr>
  <tr class="msg" id="L18"><td class="ts-cell"><a class="ts" href="#L18" alt="Tue Sep 24 2013 01:08:29 GMT-0700 (Pacific Daylight Time)">08:08</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-3" title=darobin@irc>darobin</span>&gt;</div></td><td class="msg-cell">jgraham: hey man! We&#039;ve got some cookies over here!</td></tr>
  <tr class="msg" id="L19"><td class="ts-cell"><a class="ts" href="#L19" alt="Tue Sep 24 2013 01:08:41 GMT-0700 (Pacific Daylight Time)">08:08</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-5" title=hallvors@irc>hallvors</span>&gt;</div></td><td class="msg-cell">:)</td></tr>
  <tr class="msg" id="L20"><td class="ts-cell"><a class="ts" href="#L20" alt="Tue Sep 24 2013 01:09:05 GMT-0700 (Pacific Daylight Time)">08:09</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">What&#039;s up?</td></tr>
  <tr class="msg" id="L21"><td class="ts-cell"><a class="ts" href="#L21" alt="Tue Sep 24 2013 01:11:05 GMT-0700 (Pacific Daylight Time)">08:11</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-5" title=hallvors@irc>hallvors</span>&gt;</div></td><td class="msg-cell">Ms2ger: well, busy responding to you in bug 918764 :)</td></tr>
  <tr class="msg" id="L22"><td class="ts-cell"><a class="ts" href="#L22" alt="Tue Sep 24 2013 01:11:52 GMT-0700 (Pacific Daylight Time)">08:11</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-5" title=hallvors@irc>hallvors</span>&gt;</div></td><td class="msg-cell">I&#039;m also curious what the best way to use testharness.js&#039;s assert_throws() for a TypeError is</td></tr>
  <tr class="msg" id="L23"><td class="ts-cell"><a class="ts" href="#L23" alt="Tue Sep 24 2013 01:12:07 GMT-0700 (Pacific Daylight Time)">08:12</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">assert_throws(new TypeError(), fn)</td></tr>
  <tr class="msg" id="L24"><td class="ts-cell"><a class="ts" href="#L24" alt="Tue Sep 24 2013 01:12:11 GMT-0700 (Pacific Daylight Time)">08:12</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-5" title=hallvors@irc>hallvors</span>&gt;</div></td><td class="msg-cell">assert_throws(&quot;TypeError&quot;, ... doesn&#039;t work</td></tr>
  <tr class="msg" id="L25"><td class="ts-cell"><a class="ts" href="#L25" alt="Tue Sep 24 2013 01:12:16 GMT-0700 (Pacific Daylight Time)">08:12</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-5" title=hallvors@irc>hallvors</span>&gt;</div></td><td class="msg-cell">OK</td></tr>
  <tr class="msg" id="L26"><td class="ts-cell"><a class="ts" href="#L26" alt="Tue Sep 24 2013 01:12:42 GMT-0700 (Pacific Daylight Time)">08:12</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">It&#039;s a bit strange :)</td></tr>
  <tr class="msg" id="L27"><td class="ts-cell"><a class="ts" href="#L27" alt="Tue Sep 24 2013 01:12:45 GMT-0700 (Pacific Daylight Time)">08:12</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-5" title=hallvors@irc>hallvors</span>&gt;</div></td><td class="msg-cell">I was trying to decide between new TypeError() and passing in {name:&#039;TypeError&#039; ..</td></tr>
  <tr class="msg" id="L28"><td class="ts-cell"><a class="ts" href="#L28" alt="Tue Sep 24 2013 01:13:32 GMT-0700 (Pacific Daylight Time)">08:13</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-5" title=hallvors@irc>hallvors</span>&gt;</div></td><td class="msg-cell">It&#039;s strange mainly because it&#039;s somewhat inconsistent with the assert_throws(&quot;FooError&quot; style for the DOM exceptions..</td></tr>
  <tr class="msg" id="L29"><td class="ts-cell"><a class="ts" href="#L29" alt="Tue Sep 24 2013 01:14:41 GMT-0700 (Pacific Daylight Time)">08:14</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-14" title=zcorpan@irc>zcorpan</span>&gt;</div></td><td class="msg-cell">Hixie_: <a href="http://www.whatwg.org/specs/web-apps/current-work/multipage/fragment-links.js">http://www.whatwg.org/specs/web-apps/current-work/multipage/fragment-links.js</a> seems broken</td></tr>
  <tr class="msg" id="L30"><td class="ts-cell"><a class="ts" href="#L30" alt="Tue Sep 24 2013 01:14:51 GMT-0700 (Pacific Daylight Time)">08:14</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-5" title=hallvors@irc>hallvors</span>&gt;</div></td><td class="msg-cell">I would prefer adding more known strings so that us poor test authors don&#039;t have to figure out what syntax to use for the error we&#039;re after..</td></tr>
  <tr class="msg" id="L31"><td class="ts-cell"><a class="ts" href="#L31" alt="Tue Sep 24 2013 01:15:19 GMT-0700 (Pacific Daylight Time)">08:15</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-5" title=hallvors@irc>hallvors</span>&gt;</div></td><td class="msg-cell">ah, zcorpan - perhaps I should have bothered you and not James about this ;-)</td></tr>
  <tr class="msg" id="L32"><td class="ts-cell"><a class="ts" href="#L32" alt="Tue Sep 24 2013 01:17:03 GMT-0700 (Pacific Daylight Time)">08:17</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-5" title=hallvors@irc>hallvors</span>&gt;</div></td><td class="msg-cell">anyway Ms2ger, if you have a moment please reply in bug 918764 - thx :)</td></tr>
  <tr class="msg" id="L33"><td class="ts-cell"><a class="ts" href="#L33" alt="Tue Sep 24 2013 01:17:07 GMT-0700 (Pacific Daylight Time)">08:17</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-14" title=zcorpan@irc>zcorpan</span>&gt;</div></td><td class="msg-cell">hallvors: if you can construct the real thing, do that, if not (DOMException), use a custom object</td></tr>
  <tr class="msg" id="L34"><td class="ts-cell"><a class="ts" href="#L34" alt="Tue Sep 24 2013 01:18:15 GMT-0700 (Pacific Daylight Time)">08:18</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-14" title=zcorpan@irc>zcorpan</span>&gt;</div></td><td class="msg-cell">or a string for DOMException...</td></tr>
  <tr class="msg" id="L35"><td class="ts-cell"><a class="ts" href="#L35" alt="Tue Sep 24 2013 01:18:46 GMT-0700 (Pacific Daylight Time)">08:18</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-14" title=zcorpan@irc>zcorpan</span>&gt;</div></td><td class="msg-cell">not super obvious, i agree</td></tr>
  <tr class="msg" id="L36"><td class="ts-cell"><a class="ts" href="#L36" alt="Tue Sep 24 2013 01:19:46 GMT-0700 (Pacific Daylight Time)">08:19</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-14" title=zcorpan@irc>zcorpan</span>&gt;</div></td><td class="msg-cell">Philip`: do you know anything about why <a href="http://www.whatwg.org/specs/web-apps/current-work/multipage/fragment-links.js">http://www.whatwg.org/specs/web-apps/current-work/multipage/fragment-links.js</a> is broken?</td></tr>
  <tr class="msg" id="L37"><td class="ts-cell"><a class="ts" href="#L37" alt="Tue Sep 24 2013 01:19:58 GMT-0700 (Pacific Daylight Time)">08:19</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">Don&#039;t the docs say all that, though?</td></tr>
  <tr class="msg" id="L38"><td class="ts-cell"><a class="ts" href="#L38" alt="Tue Sep 24 2013 01:20:13 GMT-0700 (Pacific Daylight Time)">08:20</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-14" title=zcorpan@irc>zcorpan</span>&gt;</div></td><td class="msg-cell">yeah</td></tr>
  <tr class="msg" id="L39"><td class="ts-cell"><a class="ts" href="#L39" alt="Tue Sep 24 2013 01:23:31 GMT-0700 (Pacific Daylight Time)">08:23</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-9" title=Philip`@irc>Philip`</span>&gt;</div></td><td class="msg-cell">zcorpan: It worked when I last looked at it, so someone must have changed something in the past five years</td></tr>
  <tr class="msg" id="L40"><td class="ts-cell"><a class="ts" href="#L40" alt="Tue Sep 24 2013 01:24:04 GMT-0700 (Pacific Daylight Time)">08:24</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-14" title=zcorpan@irc>zcorpan</span>&gt;</div></td><td class="msg-cell">helpful!</td></tr>
  <tr class="msg" id="L41"><td class="ts-cell"><a class="ts" href="#L41" alt="Tue Sep 24 2013 01:24:55 GMT-0700 (Pacific Daylight Time)">08:24</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-14" title=zcorpan@irc>zcorpan</span>&gt;</div></td><td class="msg-cell">xref relies on it in <a href="http://html.py">html.py</a> so now i can&#039;t generate html-differences</td></tr>
  <tr class="msg" id="L42"><td class="ts-cell"><a class="ts" href="#L42" alt="Tue Sep 24 2013 01:26:05 GMT-0700 (Pacific Daylight Time)">08:26</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-9" title=Philip`@irc>Philip`</span>&gt;</div></td><td class="msg-cell">zcorpan: The file looks truncated, so I guess either the upload was incomplete or the script aborted halfway through</td></tr>
  <tr class="msg" id="L43"><td class="ts-cell"><a class="ts" href="#L43" alt="Tue Sep 24 2013 01:26:18 GMT-0700 (Pacific Daylight Time)">08:26</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-9" title=Philip`@irc>Philip`</span>&gt;</div></td><td class="msg-cell">zcorpan: but I don&#039;t know where the script is run now, or how to see what errors it might have emitted</td></tr>
  <tr class="msg" id="L44"><td class="ts-cell"><a class="ts" href="#L44" alt="Tue Sep 24 2013 01:26:20 GMT-0700 (Pacific Daylight Time)">08:26</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title=hsivonen@irc>hsivonen</span>&gt;</div></td><td class="msg-cell">I wonder if these seriously bad idea encodings are truly needed for email</td></tr>
  <tr class="msg" id="L45"><td class="ts-cell"><a class="ts" href="#L45" alt="Tue Sep 24 2013 01:26:33 GMT-0700 (Pacific Daylight Time)">08:26</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title=hsivonen@irc>hsivonen</span>&gt;</div></td><td class="msg-cell">UTF-7 and its friends</td></tr>
  <tr class="msg" id="L46"><td class="ts-cell"><a class="ts" href="#L46" alt="Tue Sep 24 2013 01:26:55 GMT-0700 (Pacific Daylight Time)">08:26</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-14" title=zcorpan@irc>zcorpan</span>&gt;</div></td><td class="msg-cell">yeah failed upload seems plausible, i&#039;ll wait a bit and see if it fixes itself</td></tr>
  <tr class="msg" id="L47"><td class="ts-cell"><a class="ts" href="#L47" alt="Tue Sep 24 2013 01:28:17 GMT-0700 (Pacific Daylight Time)">08:28</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-14" title=zcorpan@irc>zcorpan</span>&gt;</div></td><td class="msg-cell">hsivonen: maybe first step towards making them not truly needed is to identify clients that send utf-7 by default, and second clients that support sending utf-7</td></tr>
  <tr class="msg" id="L48"><td class="ts-cell"><a class="ts" href="#L48" alt="Tue Sep 24 2013 01:29:10 GMT-0700 (Pacific Daylight Time)">08:29</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-14" title=zcorpan@irc>zcorpan</span>&gt;</div></td><td class="msg-cell">although i can imagine the problem isn&#039;t just with clients but with scripts that send email (like newsletters)</td></tr>
  <tr class="msg" id="L49"><td class="ts-cell"><a class="ts" href="#L49" alt="Tue Sep 24 2013 01:31:44 GMT-0700 (Pacific Daylight Time)">08:31</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title=hsivonen@irc>hsivonen</span>&gt;</div></td><td class="msg-cell">encodings are fractally wrong</td></tr>
  <tr class="msg" id="L50"><td class="ts-cell"><a class="ts" href="#L50" alt="Tue Sep 24 2013 01:31:56 GMT-0700 (Pacific Daylight Time)">08:31</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title=hsivonen@irc>hsivonen</span>&gt;</div></td><td class="msg-cell">every time I try to fix a bug, I find another bug that I should fix</td></tr>
  <tr class="msg" id="L51"><td class="ts-cell"><a class="ts" href="#L51" alt="Tue Sep 24 2013 01:32:49 GMT-0700 (Pacific Daylight Time)">08:32</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-14" title=zcorpan@irc>zcorpan</span>&gt;</div></td><td class="msg-cell"><a href="http://programmers.stackexchange.com/questions/163930/how-relevant-is-utf-7-when-it-comes-to-parsing-emails">http://programmers.stackexchange.com/questions/163930/how-relevant-is-utf-7-when-it-comes-to-parsing-emails</a></td></tr>
  <tr class="msg" id="L52"><td class="ts-cell"><a class="ts" href="#L52" alt="Tue Sep 24 2013 01:34:43 GMT-0700 (Pacific Daylight Time)">08:34</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title=hsivonen@irc>hsivonen</span>&gt;</div></td><td class="msg-cell">oh programmers.stackexchange is for conceptual questions and StackOverflow for non-conceptual ones?</td></tr>
  <tr class="msg" id="L53"><td class="ts-cell"><a class="ts" href="#L53" alt="Tue Sep 24 2013 01:35:48 GMT-0700 (Pacific Daylight Time)">08:35</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title=hsivonen@irc>hsivonen</span>&gt;</div></td><td class="msg-cell">Developing a browser engine would be easier if it wasn&#039;t used for email also.</td></tr>
  <tr class="msg" id="L54"><td class="ts-cell"><a class="ts" href="#L54" alt="Tue Sep 24 2013 01:35:58 GMT-0700 (Pacific Daylight Time)">08:35</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-14" title=zcorpan@irc>zcorpan</span>&gt;</div></td><td class="msg-cell">so hotmail bounces is utf-7 (unless they&#039;ve changed it), that&#039;s something users actually get</td></tr>
  <tr class="msg" id="L55"><td class="ts-cell"><a class="ts" href="#L55" alt="Tue Sep 24 2013 01:36:06 GMT-0700 (Pacific Daylight Time)">08:36</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title=hsivonen@irc>hsivonen</span>&gt;</div></td><td class="msg-cell">sad</td></tr>
  <tr class="msg" id="L56"><td class="ts-cell"><a class="ts" href="#L56" alt="Tue Sep 24 2013 01:39:06 GMT-0700 (Pacific Daylight Time)">08:39</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-3" title=darobin@irc>darobin</span>&gt;</div></td><td class="msg-cell">maybe it&#039;d help if we added the list of errors from <a href="http://www.ecma-international.org/ecma-262/5.1/#sec-15.11">http://www.ecma-international.org/ecma-262/5.1/#sec-15.11</a> to testharness?</td></tr>
  <tr class="msg" id="L57"><td class="ts-cell"><a class="ts" href="#L57" alt="Tue Sep 24 2013 01:40:03 GMT-0700 (Pacific Daylight Time)">08:40</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">gsnedders, merge <a href="https://github.com/html5lib/html5lib-tests/pull/4">https://github.com/html5lib/html5lib-tests/pull/4</a> already :)</td></tr>
  <tr class="msg" id="L58"><td class="ts-cell"><a class="ts" href="#L58" alt="Tue Sep 24 2013 01:41:45 GMT-0700 (Pacific Daylight Time)">08:41</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-14" title=zcorpan@irc>zcorpan</span>&gt;</div></td><td class="msg-cell">Hixie_: is fragment-links.js regenned every commit?</td></tr>
  <tr class="msg" id="L59"><td class="ts-cell"><a class="ts" href="#L59" alt="Tue Sep 24 2013 01:46:19 GMT-0700 (Pacific Daylight Time)">08:46</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-9" title=Philip`@irc>Philip`</span>&gt;</div></td><td class="msg-cell">zcorpan: It&#039;s generated by <a href="http://html5.googlecode.com/svn/trunk/spec-splitter/spec-splitter.py">http://html5.googlecode.com/svn/trunk/spec-splitter/spec-splitter.py</a> which ought to be run every commit</td></tr>
  <tr class="msg" id="L60"><td class="ts-cell"><a class="ts" href="#L60" alt="Tue Sep 24 2013 01:47:53 GMT-0700 (Pacific Daylight Time)">08:47</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-14" title=zcorpan@irc>zcorpan</span>&gt;</div></td><td class="msg-cell">ok</td></tr>
  <tr class="msg" id="L61"><td class="ts-cell"><a class="ts" href="#L61" alt="Tue Sep 24 2013 01:49:23 GMT-0700 (Pacific Daylight Time)">08:49</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-14" title=zcorpan@irc>zcorpan</span>&gt;</div></td><td class="msg-cell">the multipage version is up to date so that part still works</td></tr>
  <tr class="msg" id="L62"><td class="ts-cell"><a class="ts" href="#L62" alt="Tue Sep 24 2013 01:55:21 GMT-0700 (Pacific Daylight Time)">08:55</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title=hsivonen@irc>hsivonen</span>&gt;</div></td><td class="msg-cell">\o/ Thunderbird always sends UTF-8 to the HTML parser, so I can rip out support for mail-only encodings out of the parser</td></tr>
  <tr class="msg" id="L63"><td class="ts-cell"><a class="ts" href="#L63" alt="Tue Sep 24 2013 01:55:46 GMT-0700 (Pacific Daylight Time)">08:55</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">\o/</td></tr>
  <tr class="msg" id="L64"><td class="ts-cell"><a class="ts" href="#L64" alt="Tue Sep 24 2013 03:05:36 GMT-0700 (Pacific Daylight Time)">10:05</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title=hsivonen@irc>hsivonen</span>&gt;</div></td><td class="msg-cell">annevk: do you happen to already have test cases for the replacement encoding?</td></tr>
  <tr class="msg" id="L65"><td class="ts-cell"><a class="ts" href="#L65" alt="Tue Sep 24 2013 03:08:20 GMT-0700 (Pacific Daylight Time)">10:08</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=jgraham@irc>jgraham</span>&gt;</div></td><td class="msg-cell">hallvors: Sorry, missed the question earlier. Did you sort things out?</td></tr>
  <tr class="msg" id="L66"><td class="ts-cell"><a class="ts" href="#L66" alt="Tue Sep 24 2013 03:08:24 GMT-0700 (Pacific Daylight Time)">10:08</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title=hsivonen@irc>hsivonen</span>&gt;</div></td><td class="msg-cell">annevk: can you remind me why the spec doesn&#039;t map UTF-7 and HZ to the replacement encoding?</td></tr>
  <tr class="msg" id="L67"><td class="ts-cell"><a class="ts" href="#L67" alt="Tue Sep 24 2013 03:08:44 GMT-0700 (Pacific Daylight Time)">10:08</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=jgraham@irc>jgraham</span>&gt;</div></td><td class="msg-cell">We could probably add support for assert_throws(&quot;TypeError&quot;, function(){})</td></tr>
  <tr class="msg" id="L68"><td class="ts-cell"><a class="ts" href="#L68" alt="Tue Sep 24 2013 03:38:21 GMT-0700 (Pacific Daylight Time)">10:38</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=aleray@irc>aleray</span>&gt;</div></td><td class="msg-cell">hi, I&#039;m having problems with html5lib and lxml.etree. I don&#039;t know exactly what happens but I get a &quot;&#039;unicode&#039; object has no attribute &#039;tag&#039;&quot; AttributeError. I wrote my own serializer to tranform HTML into ConText (LaTex-like language). I have a bunch of database entries that have some fields with HTML in them. It seems to happen when the fragment of HTML I pass in is text only (not enclosed by any tag). Any Idea?</td></tr>
  <tr class="msg" id="L69"><td class="ts-cell"><a class="ts" href="#L69" alt="Tue Sep 24 2013 03:38:28 GMT-0700 (Pacific Daylight Time)">10:38</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=aleray@irc>aleray</span>&gt;</div></td><td class="msg-cell"><a href="http://dpaste.com/1394167/">http://dpaste.com/1394167/</a></td></tr>
  <tr class="msg" id="L70"><td class="ts-cell"><a class="ts" href="#L70" alt="Tue Sep 24 2013 04:06:58 GMT-0700 (Pacific Daylight Time)">11:06</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=aleray@irc>aleray</span>&gt;</div></td><td class="msg-cell">ok more info: <a href="http://dpaste.com/1394200/">http://dpaste.com/1394200/</a></td></tr>
  <tr class="msg" id="L71"><td class="ts-cell"><a class="ts" href="#L71" alt="Tue Sep 24 2013 04:07:41 GMT-0700 (Pacific Daylight Time)">11:07</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=aleray@irc>aleray</span>&gt;</div></td><td class="msg-cell">I somehow end up with a tree that contains a unicode object, and not an Element</td></tr>
  <tr class="msg" id="L72"><td class="ts-cell"><a class="ts" href="#L72" alt="Tue Sep 24 2013 04:47:02 GMT-0700 (Pacific Daylight Time)">11:47</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=jgraham@irc>jgraham</span>&gt;</div></td><td class="msg-cell">aleray: Do you have a small testcase that demonstrates the problem?</td></tr>
  <tr class="msg" id="L73"><td class="ts-cell"><a class="ts" href="#L73" alt="Tue Sep 24 2013 04:57:40 GMT-0700 (Pacific Daylight Time)">11:57</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=aleray@irc>aleray</span>&gt;</div></td><td class="msg-cell">jgraham, trying to make on</td></tr>
  <tr class="msg" id="L74"><td class="ts-cell"><a class="ts" href="#L74" alt="Tue Sep 24 2013 04:57:42 GMT-0700 (Pacific Daylight Time)">11:57</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=aleray@irc>aleray</span>&gt;</div></td><td class="msg-cell">one</td></tr>
  <tr class="msg" id="L75"><td class="ts-cell"><a class="ts" href="#L75" alt="Tue Sep 24 2013 04:57:46 GMT-0700 (Pacific Daylight Time)">11:57</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=aleray@irc>aleray</span>&gt;</div></td><td class="msg-cell">just a moment</td></tr>
  <tr class="msg" id="L76"><td class="ts-cell"><a class="ts" href="#L76" alt="Tue Sep 24 2013 05:06:27 GMT-0700 (Pacific Daylight Time)">12:06</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=aleray@irc>aleray</span>&gt;</div></td><td class="msg-cell">jgraham, here it is: <a href="http://dpaste.com/1394268/">http://dpaste.com/1394268/</a></td></tr>
  <tr class="msg" id="L77"><td class="ts-cell"><a class="ts" href="#L77" alt="Tue Sep 24 2013 05:09:09 GMT-0700 (Pacific Daylight Time)">12:09</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=aleray@irc>aleray</span>&gt;</div></td><td class="msg-cell">jgraham, works when I remove the the unicode entities</td></tr>
  <tr class="msg" id="L78"><td class="ts-cell"><a class="ts" href="#L78" alt="Tue Sep 24 2013 05:17:49 GMT-0700 (Pacific Daylight Time)">12:17</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=aleray@irc>aleray</span>&gt;</div></td><td class="msg-cell">jgraham, seems to be caused by <a href="http://lxmletree.py">lxmletree.py</a></td></tr>
  <tr class="msg" id="L79"><td class="ts-cell"><a class="ts" href="#L79" alt="Tue Sep 24 2013 05:18:00 GMT-0700 (Pacific Daylight Time)">12:18</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=jgraham@irc>jgraham</span>&gt;</div></td><td class="msg-cell">aleray: That seems like a bug in html5lib. Can you file an issue on github?</td></tr>
  <tr class="msg" id="L80"><td class="ts-cell"><a class="ts" href="#L80" alt="Tue Sep 24 2013 05:18:12 GMT-0700 (Pacific Daylight Time)">12:18</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=aleray@irc>aleray</span>&gt;</div></td><td class="msg-cell">jgraham, yes</td></tr>
  <tr class="msg" id="L81"><td class="ts-cell"><a class="ts" href="#L81" alt="Tue Sep 24 2013 05:18:29 GMT-0700 (Pacific Daylight Time)">12:18</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=jgraham@irc>jgraham</span>&gt;</div></td><td class="msg-cell">Oh, unless you have time towork out what&#039;s wrong and fix it, in which case an issue and a PR would be great ;)</td></tr>
  <tr class="msg" id="L82"><td class="ts-cell"><a class="ts" href="#L82" alt="Tue Sep 24 2013 05:18:33 GMT-0700 (Pacific Daylight Time)">12:18</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=aleray@irc>aleray</span>&gt;</div></td><td class="msg-cell">the FragmentWrapper.isstring method doesn&#039;t play nice with unicode</td></tr>
  <tr class="msg" id="L83"><td class="ts-cell"><a class="ts" href="#L83" alt="Tue Sep 24 2013 05:18:58 GMT-0700 (Pacific Daylight Time)">12:18</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=aleray@irc>aleray</span>&gt;</div></td><td class="msg-cell">jgraham, I can try, but I&#039;m always a mess with unicode stuff :)</td></tr>
  <tr class="msg" id="L84"><td class="ts-cell"><a class="ts" href="#L84" alt="Tue Sep 24 2013 05:19:49 GMT-0700 (Pacific Daylight Time)">12:19</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=jgraham@irc>jgraham</span>&gt;</div></td><td class="msg-cell">Ah, you could be right</td></tr>
  <tr class="msg" id="L85"><td class="ts-cell"><a class="ts" href="#L85" alt="Tue Sep 24 2013 05:20:18 GMT-0700 (Pacific Daylight Time)">12:20</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=jgraham@irc>jgraham</span>&gt;</div></td><td class="msg-cell">the isstring attribute looks quite python 3 ish</td></tr>
  <tr class="msg" id="L86"><td class="ts-cell"><a class="ts" href="#L86" alt="Tue Sep 24 2013 05:20:23 GMT-0700 (Pacific Daylight Time)">12:20</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=jgraham@irc>jgraham</span>&gt;</div></td><td class="msg-cell">gsnedders: &amp;</td></tr>
  <tr class="msg" id="L87"><td class="ts-cell"><a class="ts" href="#L87" alt="Tue Sep 24 2013 05:20:31 GMT-0700 (Pacific Daylight Time)">12:20</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=jgraham@irc>jgraham</span>&gt;</div></td><td class="msg-cell">s/&amp;/^/</td></tr>
  <tr class="msg" id="L88"><td class="ts-cell"><a class="ts" href="#L88" alt="Tue Sep 24 2013 05:20:49 GMT-0700 (Pacific Daylight Time)">12:20</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=aleray@irc>aleray</span>&gt;</div></td><td class="msg-cell">jgraham, basically: &quot;self.isstring = isinstance(obj, str) or isinstance(obj, bytes)&quot; should be valid for unicode type</td></tr>
  <tr class="msg" id="L89"><td class="ts-cell"><a class="ts" href="#L89" alt="Tue Sep 24 2013 05:22:59 GMT-0700 (Pacific Daylight Time)">12:22</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=jgraham@irc>jgraham</span>&gt;</div></td><td class="msg-cell">self.isstring = isinstance(obj, six.string_types)</td></tr>
  <tr class="msg" id="L90"><td class="ts-cell"><a class="ts" href="#L90" alt="Tue Sep 24 2013 05:23:08 GMT-0700 (Pacific Daylight Time)">12:23</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=jgraham@irc>jgraham</span>&gt;</div></td><td class="msg-cell">Perhaps</td></tr>
  <tr class="msg" id="L91"><td class="ts-cell"><a class="ts" href="#L91" alt="Tue Sep 24 2013 05:29:26 GMT-0700 (Pacific Daylight Time)">12:29</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=aleray@irc>aleray</span>&gt;</div></td><td class="msg-cell">jgraham, here is the relevant part of the thing: <a href="http://dpaste.com/1394290/">http://dpaste.com/1394290/</a></td></tr>
  <tr class="msg" id="L92"><td class="ts-cell"><a class="ts" href="#L92" alt="Tue Sep 24 2013 05:31:59 GMT-0700 (Pacific Daylight Time)">12:31</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=aleray@irc>aleray</span>&gt;</div></td><td class="msg-cell">sl I guess it should become &quot;self.isstring = isinstance(obj, str) or isinstance(obj, bytes) or isinstance(obj, six.text_type)&quot;</td></tr>
  <tr class="msg" id="L93"><td class="ts-cell"><a class="ts" href="#L93" alt="Tue Sep 24 2013 06:18:49 GMT-0700 (Pacific Daylight Time)">13:18</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">hsivonen: no tests, a bunch of encodings are not listed yet because I wanted to see if this was the direction we were heading</td></tr>
  <tr class="msg" id="L94"><td class="ts-cell"><a class="ts" href="#L94" alt="Tue Sep 24 2013 06:20:28 GMT-0700 (Pacific Daylight Time)">13:20</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">hsivonen: looking for the bug...</td></tr>
  <tr class="msg" id="L95"><td class="ts-cell"><a class="ts" href="#L95" alt="Tue Sep 24 2013 06:21:15 GMT-0700 (Pacific Daylight Time)">13:21</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">hsivonen: per <a href="https://www.w3.org/Bugs/Public/show_bug.cgi?id=21057">https://www.w3.org/Bugs/Public/show_bug.cgi?id=21057</a> you thought utf-7 and utf-32 might be relying on fallback</td></tr>
  <tr class="msg" id="L96"><td class="ts-cell"><a class="ts" href="#L96" alt="Tue Sep 24 2013 06:21:41 GMT-0700 (Pacific Daylight Time)">13:21</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">hsivonen: I think hz-gb-2312 is still in use</td></tr>
  <tr class="msg" id="L97"><td class="ts-cell"><a class="ts" href="#L97" alt="Tue Sep 24 2013 06:34:49 GMT-0700 (Pacific Daylight Time)">13:34</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=aleray@irc>aleray</span>&gt;</div></td><td class="msg-cell">here we go: <a href="https://github.com/html5lib/html5lib-python/issues/115">https://github.com/html5lib/html5lib-python/issues/115</a></td></tr>
  <tr class="msg" id="L98"><td class="ts-cell"><a class="ts" href="#L98" alt="Tue Sep 24 2013 06:35:04 GMT-0700 (Pacific Daylight Time)">13:35</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=jgraham@irc>jgraham</span>&gt;</div></td><td class="msg-cell">aleray: Thatnks</td></tr>
  <tr class="msg" id="L99"><td class="ts-cell"><a class="ts" href="#L99" alt="Tue Sep 24 2013 06:35:08 GMT-0700 (Pacific Daylight Time)">13:35</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=jgraham@irc>jgraham</span>&gt;</div></td><td class="msg-cell">*Thanks</td></tr>
  <tr class="msg" id="L100"><td class="ts-cell"><a class="ts" href="#L100" alt="Tue Sep 24 2013 06:35:58 GMT-0700 (Pacific Daylight Time)">13:35</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=aleray@irc>aleray</span>&gt;</div></td><td class="msg-cell">jgraham, happy to help :) I&#039;m quite in love with html5lib. I didn&#039;t wrtie any test I don&#039;t know exactly how it should be done</td></tr>
  <tr class="msg" id="L101"><td class="ts-cell"><a class="ts" href="#L101" alt="Tue Sep 24 2013 07:13:32 GMT-0700 (Pacific Daylight Time)">14:13</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">Hrm, idlharness doesn&#039;t seem to like variadic arguments</td></tr>
  <tr class="msg" id="L102"><td class="ts-cell"><a class="ts" href="#L102" alt="Tue Sep 24 2013 07:20:00 GMT-0700 (Pacific Daylight Time)">14:20</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-3" title=darobin@irc>darobin</span>&gt;</div></td><td class="msg-cell">Ms2ger: is this from an existing IDL?</td></tr>
  <tr class="msg" id="L103"><td class="ts-cell"><a class="ts" href="#L103" alt="Tue Sep 24 2013 07:20:13 GMT-0700 (Pacific Daylight Time)">14:20</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">DOMTokenList</td></tr>
  <tr class="msg" id="L104"><td class="ts-cell"><a class="ts" href="#L104" alt="Tue Sep 24 2013 07:21:48 GMT-0700 (Pacific Daylight Time)">14:21</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">annevk, r? <a href="https://critic.hoppipolla.co.uk/r/288">https://critic.hoppipolla.co.uk/r/288</a></td></tr>
  <tr class="msg" id="L105"><td class="ts-cell"><a class="ts" href="#L105" alt="Tue Sep 24 2013 07:22:55 GMT-0700 (Pacific Daylight Time)">14:22</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">r+</td></tr>
  <tr class="msg" id="L106"><td class="ts-cell"><a class="ts" href="#L106" alt="Tue Sep 24 2013 07:23:39 GMT-0700 (Pacific Daylight Time)">14:23</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">Ta</td></tr>
  <tr class="msg" id="L107"><td class="ts-cell"><a class="ts" href="#L107" alt="Tue Sep 24 2013 07:31:41 GMT-0700 (Pacific Daylight Time)">14:31</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=jgraham@irc>jgraham</span>&gt;</div></td><td class="msg-cell">gsnedders: I think the point is that in this case lxml gives unicode on py2</td></tr>
  <tr class="msg" id="L108"><td class="ts-cell"><a class="ts" href="#L108" alt="Tue Sep 24 2013 07:32:08 GMT-0700 (Pacific Daylight Time)">14:32</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=jgraham@irc>jgraham</span>&gt;</div></td><td class="msg-cell">zcorpan: I think I accidentially fixed the bug that caused reviews to not get closed on critic</td></tr>
  <tr class="msg" id="L109"><td class="ts-cell"><a class="ts" href="#L109" alt="Tue Sep 24 2013 07:32:20 GMT-0700 (Pacific Daylight Time)">14:32</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-14" title=zcorpan@irc>zcorpan</span>&gt;</div></td><td class="msg-cell">jgraham: nice</td></tr>
  <tr class="msg" id="L110"><td class="ts-cell"><a class="ts" href="#L110" alt="Tue Sep 24 2013 07:32:42 GMT-0700 (Pacific Daylight Time)">14:32</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=jgraham@irc>jgraham</span>&gt;</div></td><td class="msg-cell">In other news, I now have pywebsockets tests running on <a href="http://localhost">localhost</a> :)</td></tr>
  <tr class="msg" id="L111"><td class="ts-cell"><a class="ts" href="#L111" alt="Tue Sep 24 2013 07:33:02 GMT-0700 (Pacific Daylight Time)">14:33</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-14" title=zcorpan@irc>zcorpan</span>&gt;</div></td><td class="msg-cell">jgraham: i guess it won&#039;t close the reviews currently open with a closed PR though?</td></tr>
  <tr class="msg" id="L112"><td class="ts-cell"><a class="ts" href="#L112" alt="Tue Sep 24 2013 07:33:02 GMT-0700 (Pacific Daylight Time)">14:33</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=jgraham@irc>jgraham</span>&gt;</div></td><td class="msg-cell">Well I mean pywebsockets-based tests for the websockets api</td></tr>
  <tr class="msg" id="L113"><td class="ts-cell"><a class="ts" href="#L113" alt="Tue Sep 24 2013 07:33:12 GMT-0700 (Pacific Daylight Time)">14:33</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=jgraham@irc>jgraham</span>&gt;</div></td><td class="msg-cell">zcorpan: No</td></tr>
  <tr class="msg" id="L114"><td class="ts-cell"><a class="ts" href="#L114" alt="Tue Sep 24 2013 08:27:29 GMT-0700 (Pacific Daylight Time)">15:27</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=jgraham@irc>jgraham</span>&gt;</div></td><td class="msg-cell">hallvors: So you are working on XMLHttpRequest tests, right? :)</td></tr>
  <tr class="msg" id="L115"><td class="ts-cell"><a class="ts" href="#L115" alt="Tue Sep 24 2013 08:28:03 GMT-0700 (Pacific Daylight Time)">15:28</a></td><td class="nick-cell"><div class="m-ov"><span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span></div></td><td class="msg-cell">sees outstanding xhr reviews</td></tr>
  <tr class="msg" id="L116"><td class="ts-cell"><a class="ts" href="#L116" alt="Tue Sep 24 2013 08:28:21 GMT-0700 (Pacific Daylight Time)">15:28</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-5" title=hallvors@irc>hallvors</span>&gt;</div></td><td class="msg-cell">jgraham: yes. A little bit :)</td></tr>
  <tr class="msg" id="L117"><td class="ts-cell"><a class="ts" href="#L117" alt="Tue Sep 24 2013 08:28:33 GMT-0700 (Pacific Daylight Time)">15:28</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-5" title=hallvors@irc>hallvors</span>&gt;</div></td><td class="msg-cell">Found some issues during a Gecko test run, will fix them</td></tr>
  <tr class="msg" id="L118"><td class="ts-cell"><a class="ts" href="#L118" alt="Tue Sep 24 2013 08:29:03 GMT-0700 (Pacific Daylight Time)">15:29</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=jgraham@irc>jgraham</span>&gt;</div></td><td class="msg-cell">hallvors: You don&#039;t have some hidden desire to rewrite the PHP in python so that we can run them in automation, do you?</td></tr>
  <tr class="msg" id="L119"><td class="ts-cell"><a class="ts" href="#L119" alt="Tue Sep 24 2013 08:29:32 GMT-0700 (Pacific Daylight Time)">15:29</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-5" title=hallvors@irc>hallvors</span>&gt;</div></td><td class="msg-cell">Those PHP files are pretty simple, no?</td></tr>
  <tr class="msg" id="L120"><td class="ts-cell"><a class="ts" href="#L120" alt="Tue Sep 24 2013 08:29:51 GMT-0700 (Pacific Daylight Time)">15:29</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-5" title=hallvors@irc>hallvors</span>&gt;</div></td><td class="msg-cell">should be easy enough</td></tr>
  <tr class="msg" id="L121"><td class="ts-cell"><a class="ts" href="#L121" alt="Tue Sep 24 2013 08:30:03 GMT-0700 (Pacific Daylight Time)">15:30</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-5" title=hallvors@irc>hallvors</span>&gt;</div></td><td class="msg-cell">(aka &quot;famous last words&quot; I guess :-))</td></tr>
  <tr class="msg" id="L122"><td class="ts-cell"><a class="ts" href="#L122" alt="Tue Sep 24 2013 08:30:27 GMT-0700 (Pacific Daylight Time)">15:30</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=jgraham@irc>jgraham</span>&gt;</div></td><td class="msg-cell">I don&#039;t know, I just know that they are taking up a lot of my screen :)</td></tr>
  <tr class="msg" id="L123"><td class="ts-cell"><a class="ts" href="#L123" alt="Tue Sep 24 2013 08:30:54 GMT-0700 (Pacific Daylight Time)">15:30</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=jgraham@irc>jgraham</span>&gt;</div></td><td class="msg-cell">I haven&#039;t really looked</td></tr>
  <tr class="msg" id="L124"><td class="ts-cell"><a class="ts" href="#L124" alt="Tue Sep 24 2013 08:32:08 GMT-0700 (Pacific Daylight Time)">15:32</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-5" title=hallvors@irc>hallvors</span>&gt;</div></td><td class="msg-cell">well, some of them might be tricky if they try to do corner case stuff. On the other hand, some of those corner cases might be easier to trigger in a less hacky language ;-)</td></tr>
  <tr class="msg" id="L125"><td class="ts-cell"><a class="ts" href="#L125" alt="Tue Sep 24 2013 08:32:45 GMT-0700 (Pacific Daylight Time)">15:32</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-5" title=hallvors@irc>hallvors</span>&gt;</div></td><td class="msg-cell">although I think supporting PHP in whatever test framework you&#039;re aiming for might lower the bar to participation ;-)</td></tr>
  <tr class="msg" id="L126"><td class="ts-cell"><a class="ts" href="#L126" alt="Tue Sep 24 2013 08:33:39 GMT-0700 (Pacific Daylight Time)">15:33</a></td><td class="nick-cell"><div class="m-ov"><span class="nick nick-5" title=hallvors@irc>hallvors</span></div></td><td class="msg-cell">thinks PHP is a great language for beginners because of the excellent docs, and is prepared for being flamed &amp; lynched for saying so</td></tr>
  <tr class="msg" id="L127"><td class="ts-cell"><a class="ts" href="#L127" alt="Tue Sep 24 2013 08:35:39 GMT-0700 (Pacific Daylight Time)">15:35</a></td><td class="nick-cell"><div class="m-ov"><span class="nick nick-6" title=jgraham@irc>jgraham</span></div></td><td class="msg-cell">wonders where these excellent docs are, only found the slightly rubbish built-in ones</td></tr>
  <tr class="msg" id="L128"><td class="ts-cell"><a class="ts" href="#L128" alt="Tue Sep 24 2013 08:35:46 GMT-0700 (Pacific Daylight Time)">15:35</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=jgraham@irc>jgraham</span>&gt;</div></td><td class="msg-cell">:p</td></tr>
  <tr class="msg" id="L129"><td class="ts-cell"><a class="ts" href="#L129" alt="Tue Sep 24 2013 08:36:25 GMT-0700 (Pacific Daylight Time)">15:36</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">if you know how PHP works, it&#039;s pretty great for small stuff, otherwise many pitfalls</td></tr>
  <tr class="msg" id="L130"><td class="ts-cell"><a class="ts" href="#L130" alt="Tue Sep 24 2013 08:45:28 GMT-0700 (Pacific Daylight Time)">15:45</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=zewt@irc>zewt</span>&gt;</div></td><td class="msg-cell">probably python&#039;s biggest flaw is that its syntax sucks for php-style templating</td></tr>
  <tr class="msg" id="L131"><td class="ts-cell"><a class="ts" href="#L131" alt="Tue Sep 24 2013 09:11:40 GMT-0700 (Pacific Daylight Time)">16:11</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=gsnedders@irc>gsnedders</span>&gt;</div></td><td class="msg-cell">jgraham: WTF. It shouldn&#039;t ever, I believe.</td></tr>
  <tr class="msg" id="L132"><td class="ts-cell"><a class="ts" href="#L132" alt="Tue Sep 24 2013 09:11:47 GMT-0700 (Pacific Daylight Time)">16:11</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=gsnedders@irc>gsnedders</span>&gt;</div></td><td class="msg-cell">jgraham: Well, will look later.</td></tr>
  <tr class="msg" id="L133"><td class="ts-cell"><a class="ts" href="#L133" alt="Tue Sep 24 2013 09:12:20 GMT-0700 (Pacific Daylight Time)">16:12</a></td><td class="nick-cell"><div class="m-ov"><span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span></div></td><td class="msg-cell">pokes gsnedders about that PR</td></tr>
  <tr class="msg" id="L134"><td class="ts-cell"><a class="ts" href="#L134" alt="Tue Sep 24 2013 09:13:38 GMT-0700 (Pacific Daylight Time)">16:13</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=gsnedders@irc>gsnedders</span>&gt;</div></td><td class="msg-cell">Ms2ger: Which?</td></tr>
  <tr class="msg" id="L135"><td class="ts-cell"><a class="ts" href="#L135" alt="Tue Sep 24 2013 09:13:47 GMT-0700 (Pacific Daylight Time)">16:13</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell"><a href="https://github.com/html5lib/html5lib-tests/pull/4">https://github.com/html5lib/html5lib-tests/pull/4</a></td></tr>
  <tr class="msg" id="L136"><td class="ts-cell"><a class="ts" href="#L136" alt="Tue Sep 24 2013 09:28:00 GMT-0700 (Pacific Daylight Time)">16:28</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=jgraham@irc>jgraham</span>&gt;</div></td><td class="msg-cell">So, should I consider the possibility of browsers sending multiple headers with the same name?</td></tr>
  <tr class="msg" id="L137"><td class="ts-cell"><a class="ts" href="#L137" alt="Tue Sep 24 2013 09:39:03 GMT-0700 (Pacific Daylight Time)">16:39</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=gsnedders@irc>gsnedders</span>&gt;</div></td><td class="msg-cell">Ms2ger: Oh, okay. Then sure.</td></tr>
  <tr class="msg" id="L138"><td class="ts-cell"><a class="ts" href="#L138" alt="Tue Sep 24 2013 13:20:12 GMT-0700 (Pacific Daylight Time)">20:20</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=gsnedders@irc>gsnedders</span>&gt;</div></td><td class="msg-cell">Ms2ger: if you read logs, dealt with that PR</td></tr>
  <tr class="msg" id="L139"><td class="ts-cell"><a class="ts" href="#L139" alt="Tue Sep 24 2013 13:27:01 GMT-0700 (Pacific Daylight Time)">20:27</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=gsnedders@irc>gsnedders</span>&gt;</div></td><td class="msg-cell">Ms2ger: If you deal with <a href="https://github.com/html5lib/html5lib-tests/pull/21">https://github.com/html5lib/html5lib-tests/pull/21</a> we&#039;re even. ;P</td></tr>
  <tr class="msg" id="L140"><td class="ts-cell"><a class="ts" href="#L140" alt="Tue Sep 24 2013 13:39:19 GMT-0700 (Pacific Daylight Time)">20:39</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=Hixie_@irc>Hixie_</span>&gt;</div></td><td class="msg-cell">oh man</td></tr>
  <tr class="msg" id="L141"><td class="ts-cell"><a class="ts" href="#L141" alt="Tue Sep 24 2013 13:39:27 GMT-0700 (Pacific Daylight Time)">20:39</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=Hixie_@irc>Hixie_</span>&gt;</div></td><td class="msg-cell">setting up a new event loop is gonna suck</td></tr>
  <tr class="msg" id="L142"><td class="ts-cell"><a class="ts" href="#L142" alt="Tue Sep 24 2013 13:40:00 GMT-0700 (Pacific Daylight Time)">20:40</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=Hixie_@irc>Hixie_</span>&gt;</div></td><td class="msg-cell">(a per-tab event loop for session history, independent of the event loops of any actual tabs)</td></tr>
  <tr class="msg" id="L143"><td class="ts-cell"><a class="ts" href="#L143" alt="Tue Sep 24 2013 13:40:07 GMT-0700 (Pacific Daylight Time)">20:40</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=Hixie_@irc>Hixie_</span>&gt;</div></td><td class="msg-cell">(er, of any actual browsing contexts)</td></tr>
  <tr class="msg" id="L144"><td class="ts-cell"><a class="ts" href="#L144" alt="Tue Sep 24 2013 13:45:41 GMT-0700 (Pacific Daylight Time)">20:45</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=gsnedders@irc>gsnedders</span>&gt;</div></td><td class="msg-cell">Who deals with parser stuff at WebKit nowadays?</td></tr>
  <tr class="msg" id="L145"><td class="ts-cell"><a class="ts" href="#L145" alt="Tue Sep 24 2013 13:58:26 GMT-0700 (Pacific Daylight Time)">20:58</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-14" title=zcorpan@irc>zcorpan</span>&gt;</div></td><td class="msg-cell">heycam|away: &quot;The following extended attribute is applicable to dictionaries: [Constructor].&quot; is that right?</td></tr>
  <tr class="msg" id="L146"><td class="ts-cell"><a class="ts" href="#L146" alt="Tue Sep 24 2013 15:01:52 GMT-0700 (Pacific Daylight Time)">22:01</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=Hixie_@irc>Hixie_</span>&gt;</div></td><td class="msg-cell">tantek: microformats just honours dir=&quot;&quot; transparently, right? there&#039;s no special magic?</td></tr>
  <tr class="msg" id="L147"><td class="ts-cell"><a class="ts" href="#L147" alt="Tue Sep 24 2013 15:06:54 GMT-0700 (Pacific Daylight Time)">22:06</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-8" title=tantek@irc>tantek</span>&gt;</div></td><td class="msg-cell">Hixie - that&#039;s right</td></tr>
  <tr class="msg" id="L148"><td class="ts-cell"><a class="ts" href="#L148" alt="Tue Sep 24 2013 15:07:13 GMT-0700 (Pacific Daylight Time)">22:07</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-8" title=tantek@irc>tantek</span>&gt;</div></td><td class="msg-cell">did someone raise an issue about dir?</td></tr>
  <tr class="msg" id="L149"><td class="ts-cell"><a class="ts" href="#L149" alt="Tue Sep 24 2013 15:07:41 GMT-0700 (Pacific Daylight Time)">22:07</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=Hixie_@irc>Hixie_</span>&gt;</div></td><td class="msg-cell">only tangentially. not an issue on microformats.</td></tr>
  <tr class="msg" id="L150"><td class="ts-cell"><a class="ts" href="#L150" alt="Tue Sep 24 2013 15:44:15 GMT-0700 (Pacific Daylight Time)">22:44</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">I could also change *|*:fullscreen::backdrop to ::backdrop...</td></tr>
  <tr class="msg" id="L151"><td class="ts-cell"><a class="ts" href="#L151" alt="Tue Sep 24 2013 15:44:37 GMT-0700 (Pacific Daylight Time)">22:44</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">I don&#039;t think you&#039;d be able to observe the difference</td></tr>
  <tr class="msg" id="L152"><td class="ts-cell"><a class="ts" href="#L152" alt="Tue Sep 24 2013 15:44:41 GMT-0700 (Pacific Daylight Time)">22:44</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=Hixie_@irc>Hixie_</span>&gt;</div></td><td class="msg-cell">too late :-)</td></tr>
  <tr class="msg" id="L153"><td class="ts-cell"><a class="ts" href="#L153" alt="Tue Sep 24 2013 15:44:45 GMT-0700 (Pacific Daylight Time)">22:44</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=Hixie_@irc>Hixie_</span>&gt;</div></td><td class="msg-cell">well, not too late i guess</td></tr>
  <tr class="msg" id="L154"><td class="ts-cell"><a class="ts" href="#L154" alt="Tue Sep 24 2013 15:44:52 GMT-0700 (Pacific Daylight Time)">22:44</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=Hixie_@irc>Hixie_</span>&gt;</div></td><td class="msg-cell">but i&#039;ve taken care of it</td></tr>
  <tr class="msg" id="L155"><td class="ts-cell"><a class="ts" href="#L155" alt="Tue Sep 24 2013 15:44:59 GMT-0700 (Pacific Daylight Time)">22:44</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=Hixie_@irc>Hixie_</span>&gt;</div></td><td class="msg-cell">for &lt;dialog&gt;</td></tr>
  <tr class="msg" id="L156"><td class="ts-cell"><a class="ts" href="#L156" alt="Tue Sep 24 2013 15:45:17 GMT-0700 (Pacific Daylight Time)">22:45</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=Hixie_@irc>Hixie_</span>&gt;</div></td><td class="msg-cell">i originally thought it was just defined to cover the background</td></tr>
  <tr class="msg" id="L157"><td class="ts-cell"><a class="ts" href="#L157" alt="Tue Sep 24 2013 15:45:23 GMT-0700 (Pacific Daylight Time)">22:45</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=Hixie_@irc>Hixie_</span>&gt;</div></td><td class="msg-cell">that&#039;s why i hadn&#039;t positioned it</td></tr>
  <tr class="msg" id="L158"><td class="ts-cell"><a class="ts" href="#L158" alt="Tue Sep 24 2013 15:45:25 GMT-0700 (Pacific Daylight Time)">22:45</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-6" title=Hixie_@irc>Hixie_</span>&gt;</div></td><td class="msg-cell">but it&#039;s fine</td></tr>
  <tr class="msg" id="L159"><td class="ts-cell"><a class="ts" href="#L159" alt="Tue Sep 24 2013 15:46:56 GMT-0700 (Pacific Daylight Time)">22:46</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">fair enough, no need to touch both specs at this point indeed</td></tr>

</tbody></table></div></div></div></body>
